move BlazegraphQueryContext
